feature test the builtin error types
